Add a test case and re-organize modules slightly
